:truck: Move time_as_string to strings.py
