+++ data_safe_haven/functions/__init__.py
@@ -1,4 +1,3 @@
-from .miscellaneous import time_as_string
 from .strings import (
     alphanumeric,
     b64encode,
@@ -7,6 +6,7 @@ from .strings import (
     sanitise_sre_name,
     seeded_uuid,
     sha256hash,
+    time_as_string,
     truncate_tokens,
 )
